+++ src/pages/List/ItemListComponents/ItemList.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .itemList {
   height: 400px;
