+++ src/pages/List/SelectTextComponents/SelectText.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const SelectText = ({ list, className }) => {
+  return (
+    <>
+      <li className={className}>{list}</li>
+    </>
+  );
+};
+
+export default SelectText;
