+++ src/styles/common.scss
@@ -6,16 +6,17 @@
   font-family: 'Noto Sans KR', sans-serif;
 }
 
-.conatainer {
-  max-width: 900px;
-  width: 100%;
-  margin: 0 auto;
-  background: #f5f5f5;
+.pageContainer {
+  width: 760px;
 }
-// aside 고정 너비
-.aside {
-  width: calc(100% - 600px);
-  margin-left: 30px;
+
+.mainContainer {
+  max-width: 1060px;
+  width: 100%;
+  margin: auto;
+  margin-top: 30px;
+  display: flex;
+  justify-content: space-around;
 }
 
 // 볼드
